+++ js/game-presenter.js
@@ -2,15 +2,18 @@ import {questions, result, statistics} from './data/game-data';
 import view from './view';
 import GameModel from './data/game-model';
 import Application from './application';
+import timer from './timer/timer'
+
+let stopFn;
 
 class GamePresenter {
 
   constructor(model = GameModel) {
     this.model = model;
-    this.initialLives = GameModel.lives;
     this.timer = document.querySelector('.timer-wrapper');
     this.questionsCount = questions.length;
     this.goToResults = this.goToResults.bind(this);
+    this.tick = this.tick.bind(this);
   }
 
   switchToNext(questionNum, questionsArr) {
@@ -55,26 +58,24 @@ class GamePresenter {
     return {time: this.formatTime(time), correctAnswers: currentResult.answers, percents: successPercent};
   }
 
-  gameStart() {
+  tick() {
+    this.model.time++;
+  }
 
-    if (this.model.currentQuestion !== 0) {
-      this.model.resetState();
-    }
+  gameStart() {
+    this.model.resetState();
+
+    stopFn = timer(this.model.maxTime, this.goToResults);
+    document.body.addEventListener('timer-tick', this.tick,  false);
 
     this.timer.classList.remove('invisible');
-    // window.stopFn = window.initializeCountdown(this.model.maxTime);
-
-    document.body.addEventListener('timer-end', this.goToResults, false);
-    document.body.addEventListener('timer-tick', () => {
-      this.model.time++;
-    }, false);
-
     this.switchToNext(0, questions);
   }
 
   goToResults() {
     result.stats = this.calcStats(statistics, this.model.currentQuestion, this.initialLives, this.model.lives, this.model.time);
-    // window.stopFn();
+    stopFn();
+    document.body.removeEventListener('timer-tick', this.tick);
     Application.showStats();
   }
