--- js/time-format.js
@@ -1,9 +1,0 @@
-export default (total, passed) => {
-  const minutesLeft = Math.floor((total - passed) / 60 / 1000);
-  const secondsLeft = (total - passed - minutesLeft * 60 * 1000) / 1000;
-
-  return {
-    minutes: minutesLeft,
-    seconds: secondsLeft
-  };
-};
